refactor(docs): breadcrumbs dx
