+++ apps/docs/content/components/breadcrumbs/colors.ts
@@ -1,22 +1,4 @@
-const App = `import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
-
-export default function App() {
-  const colors = ["foreground", "primary", "secondary", "success", "warning", "danger"];
-
-  return (
-    <div className="flex flex-col flex-wrap gap-4">
-      {colors.map((color) => (
-        <Breadcrumbs key={color} color={color}>
-          <BreadcrumbItem>Home</BreadcrumbItem>
-          <BreadcrumbItem>Music</BreadcrumbItem>
-          <BreadcrumbItem>Artist</BreadcrumbItem>
-          <BreadcrumbItem>Album</BreadcrumbItem>
-          <BreadcrumbItem>Song</BreadcrumbItem>
-        </Breadcrumbs>
-      ))}
-    </div>
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
