+++ apps/docs/content/components/breadcrumbs/controlled.ts
@@ -1,54 +1,5 @@
-const App = `import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
-
-export default function App() {
-  const [currentPage, setCurrentPage] = React.useState("song");
-
-  return (
-    <Breadcrumbs underline="active" onAction={(key) => setCurrentPage(key)}>
-      <BreadcrumbItem key="home" isCurrent={currentPage === "home"}>
-        Home
-      </BreadcrumbItem>
-      <BreadcrumbItem key="music" isCurrent={currentPage === "music"}>
-        Music
-      </BreadcrumbItem>
-      <BreadcrumbItem key="artist" isCurrent={currentPage === "artist"}>
-        Artist
-      </BreadcrumbItem>
-      <BreadcrumbItem key="album" isCurrent={currentPage === "album"}>
-        Album
-      </BreadcrumbItem>
-      <BreadcrumbItem key="song" isCurrent={currentPage === "song"}>
-        Song
-      </BreadcrumbItem>
-    </Breadcrumbs>
-  );
-}`;
-
-const AppTs = `import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
-
-export default function App() {
-  const [currentPage, setCurrentPage] = React.useState<React.Key>("song");
-
-  return (
-    <Breadcrumbs underline="active" onAction={(key) => setCurrentPage(key)}>
-      <BreadcrumbItem key="home" isCurrent={currentPage === "home"}>
-        Home
-      </BreadcrumbItem>
-      <BreadcrumbItem key="music" isCurrent={currentPage === "music"}>
-        Music
-      </BreadcrumbItem>
-      <BreadcrumbItem key="artist" isCurrent={currentPage === "artist"}>
-        Artist
-      </BreadcrumbItem>
-      <BreadcrumbItem key="album" isCurrent={currentPage === "album"}>
-        Album
-      </BreadcrumbItem>
-      <BreadcrumbItem key="song" isCurrent={currentPage === "song"}>
-        Song
-      </BreadcrumbItem>
-    </Breadcrumbs>
-  );
-}`;
+import App from "./controlled.raw.jsx?raw";
+import AppTs from "./controlled.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
