+++ apps/docs/content/components/breadcrumbs/radius.ts
@@ -1,22 +1,4 @@
-const App = `import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
-
-export default function App() {
-  const radius = [ "full", "lg", "md", "sm", "none"];
-
-  return (
-    <div className="flex flex-col flex-wrap gap-4">
-      {radius.map((r) => (
-        <Breadcrumbs key={r} radius={r} variant="solid">
-          <BreadcrumbItem>Home</BreadcrumbItem>
-          <BreadcrumbItem>Music</BreadcrumbItem>
-          <BreadcrumbItem>Artist</BreadcrumbItem>
-          <BreadcrumbItem>Album</BreadcrumbItem>
-          <BreadcrumbItem>Song</BreadcrumbItem>
-        </Breadcrumbs>
-      ))}
-    </div>
-  );
-}`;
+import App from "./radius.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
