+++ apps/docs/content/components/breadcrumbs/routing.raw.jsx
@@ -1,0 +1,13 @@
+import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Breadcrumbs>
+      <BreadcrumbItem href="/docs/components/button">Button</BreadcrumbItem>
+      <BreadcrumbItem href="/docs/components/breadcrumbs">Breadcrumbs</BreadcrumbItem>
+      <BreadcrumbItem href="/docs/components/card">Card</BreadcrumbItem>
+      <BreadcrumbItem href="/docs/components/checkbox">Checkbox</BreadcrumbItem>
+      <BreadcrumbItem href="/docs/components/code">Code</BreadcrumbItem>
+    </Breadcrumbs>
+  );
+}
