+++ apps/docs/content/components/breadcrumbs/routing.ts
@@ -1,16 +1,4 @@
-const App = `import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Breadcrumbs>
-      <BreadcrumbItem href="/docs/components/button">Button</BreadcrumbItem>
-      <BreadcrumbItem href="/docs/components/breadcrumbs">Breadcrumbs</BreadcrumbItem>
-      <BreadcrumbItem href="/docs/components/card">Card</BreadcrumbItem>
-      <BreadcrumbItem href="/docs/components/checkbox">Checkbox</BreadcrumbItem>
-      <BreadcrumbItem href="/docs/components/code">Code</BreadcrumbItem>
-    </Breadcrumbs>
-  );
-}`;
+import App from "./routing.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
