+++ apps/docs/content/components/breadcrumbs/separator.ts
@@ -1,21 +1,4 @@
-const App = `import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Breadcrumbs
-      separator="/"
-      itemClasses={{
-        separator: "px-2"
-      }}
-    >
-      <BreadcrumbItem>Home</BreadcrumbItem>
-      <BreadcrumbItem>Music</BreadcrumbItem>
-      <BreadcrumbItem>Artist</BreadcrumbItem>
-      <BreadcrumbItem>Album</BreadcrumbItem>
-      <BreadcrumbItem>Song</BreadcrumbItem>
-    </Breadcrumbs>
-  );
-}`;
+import App from "./separator.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
