+++ apps/docs/content/components/breadcrumbs/sizes.ts
@@ -1,22 +1,4 @@
-const App = `import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
-
-export default function App() {
-  const sizes = ["sm", "md", "lg"];
-
-  return (
-    <div className="flex flex-col flex-wrap gap-4">
-      {sizes.map((size) => (
-        <Breadcrumbs key={size} size={size}>
-          <BreadcrumbItem>Home</BreadcrumbItem>
-          <BreadcrumbItem>Music</BreadcrumbItem>
-          <BreadcrumbItem>Artist</BreadcrumbItem>
-          <BreadcrumbItem>Album</BreadcrumbItem>
-          <BreadcrumbItem>Song</BreadcrumbItem>
-        </Breadcrumbs>
-      ))}
-    </div>
-  );
-}`;
+import App from "./sizes.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
