+++ apps/docs/content/components/breadcrumbs/variants.ts
@@ -1,22 +1,4 @@
-const App = `import {Breadcrumbs, BreadcrumbItem} from "@nextui-org/react";
-
-export default function App() {
-  const variants = ["solid", "bordered", "light"];
-
-  return (
-    <div className="flex flex-col flex-wrap gap-4">
-      {variants.map((variant) => (
-        <Breadcrumbs key={variant} variant={variant}>
-          <BreadcrumbItem>Home</BreadcrumbItem>
-          <BreadcrumbItem>Music</BreadcrumbItem>
-          <BreadcrumbItem>Artist</BreadcrumbItem>
-          <BreadcrumbItem>Album</BreadcrumbItem>
-          <BreadcrumbItem>Song</BreadcrumbItem>
-        </Breadcrumbs>
-      ))}
-    </div>
-  );
-}`;
+import App from "./variants.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
